Added YAML support for the CLI
Fixed #67
