+++ packages/oasgraph-cli/lib/oasgraph.js
@@ -7,6 +7,7 @@ const oasgraph_1 = require("oasgraph");
 const path = require("path");
 const request = require("request");
 const fs = require("fs");
+const yaml = require("js-yaml");
 const graphql_1 = require("graphql");
 var program = require('commander');
 const app = express();
@@ -36,8 +37,13 @@ if (program.port) {
 }
 // check if the file exists 
 if (fs.existsSync(path.resolve(filePath))) {
-    let oas = require(path.resolve(filePath));
-    startGraphQLServer(oas, portNumber);
+    try {
+        let oas = readFile(path.resolve(filePath));
+        startGraphQLServer(oas, portNumber);
+    }
+    catch (e) {
+        console.error(e);
+    }
 }
 else {
     // falls back to a remote location
@@ -48,6 +54,28 @@ else {
     }
     else {
         console.error(`OASGraph reading local file error. File ${filePath} does not exist.`);
+    }
+}
+/**
+* Returns content of read JSON/YAML file.
+*
+* @param  {String} path Path to file to read
+* @return {Object}      Content of read file
+*/
+function readFile(path) {
+    try {
+        let doc;
+        if (/json$/.test(path)) {
+            doc = JSON.parse(fs.readFileSync(path, 'utf8'));
+        }
+        else if (/yaml$|yml$/.test(path)) {
+            doc = yaml.safeLoad(fs.readFileSync(path, 'utf8'));
+        }
+        return doc;
+    }
+    catch (e) {
+        console.error('Error: failed to parse YAML/JSON: ' + e);
+        return null;
     }
 }
 /**
